+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 1.0 branch
+### 1.0.2-beta3
+* added `--debug` options for `FilesBackupCommand`.
+
 ### 1.0.1-beta2
 * added `include` option for `FilesBackup` instances and `--include` option for
     `FilesBackupCommand`;